Move overlay to new package
