+++ app/src/main/java/com/github/damontecres/wholphin/ui/playback/overlay/PauseIndicator.kt
@@ -1,4 +1,4 @@
-package com.github.damontecres.wholphin.ui.playback
+package com.github.damontecres.wholphin.ui.playback.overlay
 
 import androidx.annotation.OptIn
 import androidx.compose.animation.AnimatedVisibility
