+++ app/src/main/java/com/github/damontecres/wholphin/ui/playback/overlay/SeekPreviewImage.kt
@@ -1,4 +1,4 @@
-package com.github.damontecres.wholphin.ui.playback
+package com.github.damontecres.wholphin.ui.playback.overlay
 
 import androidx.annotation.FloatRange
 import androidx.compose.foundation.Canvas
